+++ Source/WaveformDisplay.cpp
@@ -48,7 +48,7 @@ void WaveformDisplay::paint (juce::Graphics& g)
     }
     else
     {
-        g.setFont(getHeight()/4);
+        g.setFont((float)getHeight()/4);
         g.drawText("File not loaded...", getLocalBounds(),
                     juce::Justification::centred, true);   // draw some placeholder text
     }
